+++ client-quota/src/components/common/FormField.jsx
@@ -1,8 +1,9 @@
 import styled from 'styled-components';
 
-import InputSubTitle from './InputSubTitle';
+import InputSubTitle from '../reservation/createReservation/ui/InputSubTitle';
 import InputText from '@/components/common/InputText';
 
+// input text의 제목, 입력란 밑의 설명이 포함된 컴포넌트.
 const FormField = ({value, onChange, title, name, placeholder, description}) => {
     return (
         <FieldContainer>
